feat(style): inline default style within the JS library
This commit moves the default styling created by @LukyVj inside the
places.js library.

It also uses em units so that the height, margins, borders, shadows will
all be computed given the initial font-size.

I removed any very demo specific styling and moved it to the _inputs.scss
file.

We may need to tune again of course this default styling but here is a
good start.