feat(autocomplete): provide a change event on the places instance
The places instance is now an EventEmitter that will emit a single
"change" event anytime the user selects a value in the autocomplete
flow. (selected or autocompleted).

We will then wait for feedback before implementing things like
`results` events that would be handy to show suggestions on a map for
instance.